Refactor text handling: improve formatting consistency, remove unnecessary type conversions, and optimize string operations across multiple modules.

+++ src/main/kotlin/io/docops/docopsextensionssupport/button/shape/Rectangle.kt
@@ -194,7 +194,7 @@ class Rectangle(buttons: Buttons) : Regular(buttons) {
             """.trimIndent()
             }
             sb.append(chipHtml)
-            currentX += (textWidth + 8).toInt()
+            currentX += (textWidth + 8)
         }
 
         return sb.toString()
